backend/backend/requirements.txt
Django==3.2.16
django_filter==22.1
djangorestframework==3.14.0
djoser==2.2.3
python-dotenv==1.0.1
shortuuid==1.0.13
